+++ modules/clients/repositories/clients-repository.ts
@@ -1,5 +1,6 @@
 export {
   createClient,
+  deleteClient,
   getClientById,
   loadClients,
   updateClientProfile,